Updated Vulkan bindings to target the 1.4.341.1 SDK version. Improved the bindings generation for Vulkan.

+++ Source/Bindings/Massini.Bindings.Vulkan/Interop/VkExportMemoryWin32HandleInfoKHR.cs
@@ -11,9 +11,9 @@ namespace Massini.Bindings.Vulkan
         public _SECURITY_ATTRIBUTES* pAttributes;
 
         [NativeTypeName("DWORD")]
-        public uint dwAccess;
+        public nuint dwAccess;
 
         [NativeTypeName("LPCWSTR")]
-        public ushort* name;
+        public uint* name;
     }
 }
