+++ Source/Bindings/Massini.Bindings.Vulkan/Interop/VkExportMemoryWin32HandleInfoNV.cs
@@ -11,6 +11,6 @@ namespace Massini.Bindings.Vulkan
         public _SECURITY_ATTRIBUTES* pAttributes;
 
         [NativeTypeName("DWORD")]
-        public uint dwAccess;
+        public nuint dwAccess;
     }
 }
