+++ Source/Bindings/Massini.Bindings.Vulkan/Interop/VkImportFenceWin32HandleInfoKHR.cs
@@ -19,6 +19,6 @@ namespace Massini.Bindings.Vulkan
         public void* handle;
 
         [NativeTypeName("LPCWSTR")]
-        public ushort* name;
+        public uint* name;
     }
 }
